requirements.txt
pyside6==6.10.2
pyautogui==0.9.54
pyaudio==0.2.14
SpeechRecognition==3.14.5
opencv-python==4.12.0.88
easyocr==1.7.2
google-genai==1.69.0
pillow==12.1.0
python-dotenv==1.2.1
numpy==2.2.6
keyboard==0.13.5
pyperclip==1.11.0
mss==9.0.1
pydantic==2.11.5
pyinstaller==6.18.0
psutil==7.2.2
winsdk==1.0.0b10
imagehash==4.3.2
uiautomation==2.0.29
rapidocr-onnxruntime==1.4.4
pytest==9.0.2
websockets==16.0